Se crea servicio estado del monopatin

+++ SistemaMonopatines/Tarifa/src/main/java/org/example/tarifa/dto/request/TarifaRequestDTO.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Setter
 public class TarifaRequestDTO {
 
-    private Integer monto;
+    private Double monto;
     private Long viajeId;
     private Long usuarioId;
